+++ st2common/st2common/validators/api/reactor.py
@@ -15,8 +15,11 @@
 
 import six
 
+from apscheduler.triggers.cron import CronTrigger
+
 from st2common.exceptions.apivalidation import ValueValidationException
 from st2common.constants.triggers import SYSTEM_TRIGGER_TYPES
+from st2common.constants.triggers import CRON_TIMER_TRIGGER_REF
 from st2common.util import schema as util_schema
 import st2common.operators as criteria_operators
 
@@ -70,4 +73,13 @@ def validate_trigger_parameters(trigger_type_ref, parameters):
     cleaned = util_schema.validate(instance=parameters, schema=parameters_schema,
                                    cls=util_schema.CustomValidator, use_default=True,
                                    allow_default_none=True)
+
+    # Additional validation for CronTimer trigger
+    # TODO: If we need to add more checks like this we should consider abstracting this out.
+    if trigger_type_ref == CRON_TIMER_TRIGGER_REF:
+        # Validate that the user provided parameters are valid. This is required since JSON schema
+        # allows arbitrary strings, but not any arbitrary string is a valid CronTrigger argument
+        # Note: Constructor throws ValueError on invalid parameters
+        CronTrigger(**parameters)
+
     return cleaned