+++ st2common/tests/unit/test_operators.py
@@ -45,6 +45,31 @@ class OperatorTest(unittest2.TestCase):
         string = 'foo\r\nponies\nbar\nfooooo'
         self.assertTrue(op(string, '.*ponies.*'), 'Failed matchregex.')
 
+    def test_iregex(self):
+        op = operators.get_operator('iregex')
+        self.assertTrue(op('V1', 'v1$'), 'Failed iregex.')
+
+        string = 'fooPONIESbarfooooo'
+        self.assertTrue(op(string, 'ponies'), 'Failed iregex.')
+
+    def test_iregex_fail(self):
+        op = operators.get_operator('iregex')
+        self.assertFalse(op('V1_foo', 'v1$'), 'Passed iregex.')
+
+    def test_regex(self):
+        op = operators.get_operator('regex')
+        self.assertTrue(op('v1', 'v1$'), 'Failed regex.')
+
+        string = 'fooponiesbarfooooo'
+        self.assertTrue(op(string, 'ponies'), 'Failed regex.')
+
+    def test_regex_fail(self):
+        op = operators.get_operator('regex')
+        self.assertFalse(op('v1_foo', 'v1$'), 'Passed regex.')
+
+        string = 'fooPONIESbarfooooo'
+        self.assertFalse(op(string, 'ponies'), 'Passed regex.')
+
     def test_matchregex_case_variants(self):
         op = operators.get_operator('MATCHREGEX')
         self.assertTrue(op('v1', 'v1$'), 'Failed matchregex.')